move flash module to separate folder, make sequential-storage optional
